Create ArraySlices struct as interface to Array
An ArraySlices struct wraps a Blitz Array and exposes access to the Array
through a vector of Array slices (proxies to the underlying data from the main
Array).  An ArraySlices instance may be constructed by either:
    a) specifying the step size to use
    b) specifying a Blitz Array containing the sizes of slices to create
In the case of (b), the sum of all slice sizes must equal the main Array size.